requirements.txt
fastapi
uvicorn
python-multipart
olefile
reportlab
Pillow
python-docx  # <--- 워드(DOCX) 지원용 추가
